+++ wandb/sdk/sweeps/__init__.py
@@ -1,1 +1,2 @@
-from wandb.sdk.sweeps.run_state import RunState as RunState
+from wandb.sdk.sweeps.errors import SweepNotFoundError
+from wandb.sdk.sweeps.run_state import RunState
