+++ wandb/sdk/sweeps/errors.py
@@ -1,0 +1,2 @@
+class SweepNotFoundError(Exception):
+    """Raised when a sweep is not found, typically because it was deleted."""
